requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
scikit-learn>=0.24.0
networkx>=2.6.0
gymnasium>=0.26.0
stable-baselines3>=2.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tensorboard>=2.7.0
openpyxl>=3.0.0